+++ src/components/ItemCard/ItemCard.test.js
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { store } from '../../app/store';
 import ItemCard from './ItemCard';
-import ItemCardSection from './ItemCardSection';
+import ItemCardSection from './ItemCardSection'
 
 test('Sign In Page Renders Correctly', () => {
   render(
